seo_automator/requirements.txt
requests
python-dotenv
pandas
slugify
boto3 